requirements.txt
tensorflow-macos==2.12.0; sys_platform == 'darwin'  # For macOS
tensorflow-metal==0.8.0; sys_platform == 'darwin'  # Metal plugin for macOS
transformers>=4.30.0
numpy>=1.22,<1.24
huggingface-hub
pandas>=2.0.0